style: migrate style for Nav
