+++ src/route/basic-layout.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from "react-router-dom";
 
 import Footer from "../component/footer/footer.component";
-import Nav from "../component/nav/nav.component";
+import Nav from "../component/nav.component";
 
 export default function BasicLayout() {
   return (
